+++ src/main/java/com/hospital/entity/PatientCard.java
@@ -1,6 +1,8 @@
 package com.hospital.entity;
 
-import lombok.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import javax.persistence.*;
 import java.util.Set;
@@ -13,7 +15,7 @@ import java.util.Set;
 public class PatientCard {
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
 
     @OneToOne(fetch = FetchType.EAGER)
     private Client client;
